Clothing Brand Website
In this project, we designed a simple website for a clothing brand that we created. We used HTML5, CSS3 programming languages and Bootstrap Tools while developing our project. We have divided our website into three as home page, products and communication area. We have included product images, product price and product detail information about the product to be sold. There is information about our brand on the contact page.